hackathon/requirements.txt
anthropic==0.97.0
cognee==1.0.3
# Cognee 1.0.3 calls starlette.status.HTTP_422_UNPROCESSABLE_CONTENT (added in starlette 0.49)
starlette>=0.49
daytona==0.169.0
streamlit==1.56.0
pydantic>=2.0,<3
python-dotenv>=1.0,<2
pytest>=8.0,<9
